solver: move AddBuildConfig into llbsolver package
The data structures used by llbsolver shouldn't be bundled into the same
package as the actual creation of the build config.

As it was, it wasn't possible to use the data structures (in tests,
external libraries using buildkit, etc) without pulling in more of
buildkit than was probably intended.
